Adjusted Display code to make it not use any tft display code when no TFT display is defined
This is needed to allow JTAG esp-prog debugging possible as the TFT display used now (TTGO) is using GPIO 14 which is conflicting with the esp-prog debugger which is also using the same port.
When debugging with esp-prog use an esp32 without display to avoid conflicts with the JTAG debugger! (Or make sure GPIO 12 - GPIO 15 are not in use by the code already!)